+++ app/internal/modules/catalog/usecase/admin-api/category_image_interface.go
@@ -10,8 +10,8 @@ import (
 type (
 	CategoryImageService interface {
 		// GetFile - WARNING you don't forget to call item.Body.Close()
-		GetFile(ctx context.Context, categoryID mrtype.KeyInt32) (mrtype.File, error)
-		StoreFile(ctx context.Context, categoryID mrtype.KeyInt32, image mrtype.File) error
+		GetFile(ctx context.Context, categoryID mrtype.KeyInt32) (mrtype.Image, error)
+		StoreFile(ctx context.Context, categoryID mrtype.KeyInt32, image mrtype.Image) error
 		RemoveFile(ctx context.Context, categoryID mrtype.KeyInt32) error
 	}